Sort document and login
