+++ credential-exchange-types/src/format/login.rs
@@ -4,11 +4,44 @@
 
 use serde::{Deserialize, Serialize};
 
-use crate::{
-    b64url::B32,
-    format::{EditableField, Fido2Extensions},
-    B64Url, Uri,
-};
+use crate::{b64url::B32, format::EditableField, B64Url, Uri};
+
+/// A [ApiKeyCredential] contains information to interact with an Application's Programming
+/// Interface (API).
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ApiKeyCredential {
+    /// This REQUIRED member denotes the key to communicate with the API. Its internal fieldType
+    /// SHOULD be of type ConcealedString.
+    key: EditableField,
+
+    /// This OPTIONAL member denotes the username associated with the key and its internal
+    /// fieldType SHOULD be of type string
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    username: Option<EditableField>,
+
+    /// This OPTIONAL member denotes the type of the API key, such as bearer token or
+    /// JSON Web Token. It is flexible to allow any type and not restrict it to a set list of
+    /// types. Its internal fieldType SHOULD be of type string.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    key_type: Option<EditableField>,
+
+    /// This OPTIONAL member denotes the url the API key is used with and SHOULD conform to the
+    /// [URL Standard](https://url.spec.whatwg.org/). Its internal fieldType SHOULD be of type
+    /// string.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    url: Option<EditableField>,
+
+    /// This OPTIONAL member denotes the date the API key is valid from and its internal fieldType
+    /// SHOULD be of type date.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    valid_from: Option<EditableField>,
+
+    /// This OPTIONAL member denotes the date on which the API key expires
+    /// and its internal fieldType SHOULD be of type date.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    expiry_date: Option<EditableField>,
+}
 
 /// A [BasicAuthCredential] contains a username/password login credential.
 /// Can either represent a [Basic access authentication](https://www.rfc-editor.org/rfc/rfc7617)
@@ -26,72 +59,47 @@ pub struct BasicAuthCredential {
     pub password: Option<EditableField>,
 }
 
-/// Passkey
+/// A [GeneratedPasswordCredential] type represents a credential consisting of a machine-generated
+/// password.
 ///
-/// Note: Passkeys using a non-zero signature counter MUST be excluded from the export and the
-/// exporter SHOULD inform the user that such passkeys are excluded from the export. Importers MUST
-/// set a zero value for the imported passkey signature counters and MUST NOT increment them after
-/// the fact.
+/// Note: A [GeneratedPasswordCredential] is used when a password is generated independently of
+/// creating a new [BasicAuthCredential]. Some providers may offer a dedicated password generator
+/// feature. In such cases, the provider may create [GeneratedPasswordCredential] instances as
+/// deemed appropriate for the use of this feature.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct PasskeyCredential {
-    /// This member contains a [WebAuthn](https://www.w3.org/TR/webauthn-3)
-    /// [Credential ID](https://www.w3.org/TR/webauthn-3/#credential-id) which uniquely identifies
-    /// the passkey instance. The decoded raw value MUST be equal to the value given in
-    /// [PublicKeyCredential](https://www.w3.org/TR/webauthn-3/#iface-pkcredential)'s
-    /// [rawId](https://www.w3.org/TR/webauthn-3/#dom-publickeycredential-rawid) field during
-    /// [registration](https://www.w3.org/TR/webauthn-3/#registration).
-    pub credential_id: B64Url,
-    /// This member specifies the [WebAuthn](https://www.w3.org/TR/webauthn-3)
-    /// [Relying Party Identifier](https://www.w3.org/TR/webauthn-3/#relying-party-identifier) to
-    /// which the passkey instance is tied to. The value MUST be equal to the
-    /// [RP ID](https://www.w3.org/TR/webauthn-3/#rp-id) that was defined by the authenticator
-    /// during credential [registration](https://www.w3.org/TR/webauthn-3/#registration).
-    pub rp_id: String,
-    /// This member contains a [human-palatable](https://www.w3.org/TR/webauthn-3/#human-palatability)
-    /// identifier for the [user account](https://www.w3.org/TR/webauthn-3/#user-account) to which
-    /// the passkey instance is tied to. The value SHOULD be equal to the value in
-    /// [PublicKeyCredentialUserEntity](https://www.w3.org/TR/webauthn-3/#dictdef-publickeycredentialuserentity)'s
-    /// [name](https://www.w3.org/TR/webauthn-3/#dom-publickeycredentialentity-name) member given
-    /// to the authenticator during [registration](https://www.w3.org/TR/webauthn-3/#registration).
-    ///
-    /// The only case where the value MAY not be the one set during [registration](https://www.w3.org/TR/webauthn-3/#registration)
-    /// is if the [exporting provider](https://fidoalliance.org/specs/cx/cxp-v1.0-wd-20241003.html#exporting-provider)
-    /// allows the user to edit their username. In such a case, the value of
-    /// this field MUST be the user edited value. See [§ 3.3.3.1 Editability of passkey fields](https://fidoalliance.org/specs/cx/cxf-v1.0-wd-20241003.html#sctn-editability-of-passkey-fields)
-    /// for more details.
-    pub user_name: String,
-    /// This member contains a [human-palatable](https://www.w3.org/TR/webauthn-3/#human-palatability)
-    /// identifier for the [user account](https://www.w3.org/TR/webauthn-3/#user-account), intended
-    /// only for display. The value SHOULD be equal to the value in
-    /// [PublicKeyCredentialUserEntity](https://www.w3.org/TR/webauthn-3/#dictdef-publickeycredentialuserentity)'s
-    /// [displayName](https://www.w3.org/TR/webauthn-3/#dom-publickeycredentialuserentity-displayname)
-    /// member given to the authenticator during [registration](https://www.w3.org/TR/webauthn-3/#registration).
-    ///
-    /// The only case where the value MAY not be the one set during [registration](https://www.w3.org/TR/webauthn-3/#registration)
-    /// is if the [exporting provider](https://fidoalliance.org/specs/cx/cxp-v1.0-wd-20241003.html#exporting-provider)
-    /// allows the user to edit their username. In such a case, the value of
-    /// this field MUST be the user edited value. See [§ 3.3.3.1 Editability of passkey fields](https://fidoalliance.org/specs/cx/cxf-v1.0-wd-20241003.html#sctn-editability-of-passkey-fields)
-    /// for more details.
-    pub user_display_name: String,
-    /// This member contains the [user handle](https://www.w3.org/TR/webauthn-3/#user-handle) which
-    /// is the value used to identify the [user account](https://www.w3.org/TR/webauthn-3/#user-account)
-    /// associated to this passkey instance. The value MUST be equal to the value in
-    /// [PublicKeyCredentialUserEntity](https://www.w3.org/TR/webauthn-3/#dictdef-publickeycredentialuserentity)'s
-    /// [id](https://www.w3.org/TR/webauthn-3/#dom-publickeycredentialuserentity-id) member given
-    /// to the authenticator during [registration](https://www.w3.org/TR/webauthn-3/#registration).
-    pub user_handle: B64Url,
-    /// The [private key](https://www.w3.org/TR/webauthn-3/#credential-private-key) associated to
-    /// this passkey instance. The value MUST be [PKCS#8](https://www.rfc-editor.org/rfc/rfc5958)
-    /// [ASN.1 DER](https://fidoalliance.org/specs/cx/cxf-v1.0-wd-20241003.html#biblio-itu-x690-2008)
-    /// formatted byte string which is then [Base64url encoded](https://www.rfc-editor.org/rfc/rfc4648#section-5).
-    /// The value MUST give the same [public key](https://www.w3.org/TR/webauthn-3/#credential-public-key)
-    /// value that was provided by the original authenticator during [registration](https://www.w3.org/TR/webauthn-3/#registration).
-    pub key: B64Url,
-    /// This OPTIONAL member denotes the WebAuthn or CTAP2 extensions that are associated to this
-    /// passkey instance.
+pub struct GeneratedPasswordCredential {
+    /// The machine-generated password.
+    password: String,
+}
+
+/// An [SshKeyCredential] represents an SSH (Secure Shell) key pair.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SshKeyCredential {
+    /// The type of SSH key algorithm used. Common values include "ssh-rsa", "ssh-ed25519", or
+    /// "ecdsa-sha2-nistp256". This MUST be a string value representing a valid SSH public key
+    /// algorithm as defined in IANA SSH Protocol Parameters.
+    key_type: String,
+    /// The private part of the SSH key pair. This MUST be a PKCS#8 ASN.1 DER formatted byte string
+    /// which is then Base64url encoded.
+    private_key: B64Url,
+    /// This OPTIONAL member contains a user-defined string to identify or describe the key.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub fido2_extensions: Option<Fido2Extensions>,
+    key_comment: Option<String>,
+    /// This OPTIONAL member indicates when the key was created. When present, its internal
+    /// fieldType SHOULD be of type date.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    creation_date: Option<EditableField>,
+    /// This OPTIONAL member indicates when the key will expire, if applicable. When present, its
+    /// internal fieldType SHOULD be of type date.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    expiration_date: Option<EditableField>,
+    /// This OPTIONAL member indicates where the key was originally generated. E.g.,
+    /// `https://github.com/settings/ssh/new` for GitHub. When present, its internal fieldType
+    /// SHOULD be of type string.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    key_generation_source: Option<EditableField>,
 }
 
 /// Note: Enrollment in TOTP credentials historically has been quite non-standardized but typically
@@ -141,84 +149,4 @@ pub enum OTPHashAlgorithm {
     Sha512,
     #[serde(untagged)]
     Unknown(String),
-}
-
-/// An [SshKeyCredential] represents an SSH (Secure Shell) key pair.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SshKeyCredential {
-    /// The type of SSH key algorithm used. Common values include "ssh-rsa", "ssh-ed25519", or
-    /// "ecdsa-sha2-nistp256". This MUST be a string value representing a valid SSH public key
-    /// algorithm as defined in IANA SSH Protocol Parameters.
-    key_type: String,
-    /// The private part of the SSH key pair. This MUST be a PKCS#8 ASN.1 DER formatted byte string
-    /// which is then Base64url encoded.
-    private_key: B64Url,
-    /// This OPTIONAL member contains a user-defined string to identify or describe the key.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    key_comment: Option<String>,
-    /// This OPTIONAL member indicates when the key was created. When present, its internal
-    /// fieldType SHOULD be of type date.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    creation_date: Option<EditableField>,
-    /// This OPTIONAL member indicates when the key will expire, if applicable. When present, its
-    /// internal fieldType SHOULD be of type date.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    expiration_date: Option<EditableField>,
-    /// This OPTIONAL member indicates where the key was originally generated. E.g.,
-    /// `https://github.com/settings/ssh/new` for GitHub. When present, its internal fieldType
-    /// SHOULD be of type string.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    key_generation_source: Option<EditableField>,
-}
-
-/// A [ApiKeyCredential] contains information to interact with an Application's Programming
-/// Interface (API).
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiKeyCredential {
-    /// This REQUIRED member denotes the key to communicate with the API. Its internal fieldType
-    /// SHOULD be of type ConcealedString.
-    key: EditableField,
-
-    /// This OPTIONAL member denotes the username associated with the key and its internal
-    /// fieldType SHOULD be of type string
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    username: Option<EditableField>,
-
-    /// This OPTIONAL member denotes the type of the API key, such as bearer token or
-    /// JSON Web Token. It is flexible to allow any type and not restrict it to a set list of
-    /// types. Its internal fieldType SHOULD be of type string.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    key_type: Option<EditableField>,
-
-    /// This OPTIONAL member denotes the url the API key is used with and SHOULD conform to the
-    /// [URL Standard](https://url.spec.whatwg.org/). Its internal fieldType SHOULD be of type
-    /// string.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    url: Option<EditableField>,
-
-    /// This OPTIONAL member denotes the date the API key is valid from and its internal fieldType
-    /// SHOULD be of type date.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    valid_from: Option<EditableField>,
-
-    /// This OPTIONAL member denotes the date on which the API key expires
-    /// and its internal fieldType SHOULD be of type date.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    expiry_date: Option<EditableField>,
-}
-
-/// A [GeneratedPasswordCredential] type represents a credential consisting of a machine-generated
-/// password.
-///
-/// Note: A [GeneratedPasswordCredential] is used when a password is generated independently of
-/// creating a new [BasicAuthCredential]. Some providers may offer a dedicated password generator
-/// feature. In such cases, the provider may create [GeneratedPasswordCredential] instances as
-/// deemed appropriate for the use of this feature.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GeneratedPasswordCredential {
-    /// The machine-generated password.
-    password: String,
 }
